# CityPharmacyChain.Api.Client/Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CityPharmacyChain.Api.Host",
    "version": "1.0"
  },
  "paths": {
    "/PharmaceuticalGroup": {
      "get": {
        "tags": [
          "PharmaceuticalGroup"
        ],
        "summary": "GET запрос по получению всех объектов класса фармацевтическая группа из базы данных",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PharmaceuticalGroupFullDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PharmaceuticalGroupFullDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PharmaceuticalGroupFullDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "PharmaceuticalGroup"
        ],
        "summary": "POST запрос по добавлению объекта класса фармацевтическая группа в базу данных",
        "requestBody": {
          "description": "Объект класса фармацевтическая группа в формате JSON",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PharmaceuticalGroupDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PharmaceuticalGroupDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PharmaceuticalGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PharmaceuticalGroupFullDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PharmaceuticalGroupFullDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PharmaceuticalGroupFullDto"
                }
              }
            }
          }
        }
      }
    },
    "/PharmaceuticalGroup/{id}": {
      "get": {
        "tags": [
          "PharmaceuticalGroup"
        ],
        "summary": "GET запрос по получению объекта класса фармацевтическая группа из базы данных по его идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор фармацевтической группы",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PharmaceuticalGroupDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PharmaceuticalGroupDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PharmaceuticalGroupDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "PharmaceuticalGroup"
        ],
        "summary": "PUT запрос по модификации объекта класса фармацевтическая группа из базы данных по его идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор фармацевтической группы",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Объект класса фармацевтическая группа в формате JSON",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PharmaceuticalGroupDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PharmaceuticalGroupDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PharmaceuticalGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PharmaceuticalGroupFullDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PharmaceuticalGroupFullDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PharmaceuticalGroupFullDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "PharmaceuticalGroup"
        ],
        "summary": "DELETE запрос по удалению существующего объекта класса фармацевтическая группа из базы данных по его идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор фармацевтической группы",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "PharmaceuticalGroupDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "description": "Идентификатор препарата",
            "format": "int32"
          },
          "name": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "Название фармацевтической группы",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Класс DTO для сущности класса фармацевтическая группа"
      },
      "PharmaceuticalGroupFullDto": {
        "type": "object",
        "properties": {
          "pharmaceuticalGroupId": {
            "type": "integer",
            "description": "Идентификатор фармацевтической группы",
            "format": "int32"
          },
          "productId": {
            "type": "integer",
            "description": "Идентификатор препарата",
            "format": "int32"
          },
          "name": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "Название фармацевтической группы",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Класс DTO для сущности класса фармацевтическая группа, содержащий её идентификатор"
      },
...
    }
  }
}